release: 1.6.1 legacy JT C++ promotion and release alignment
- JT4, JT9 and JT65 active runtime paths now run through the maintained
  native C++ decode stack (FanoSequentialDecoder, JT4Decoder, JT9NarrowDecoder)
- Added FftCompat bridge replacing legacy Fortran four2a_ calls
- Q65 and WSPR are now Fortran-free
- FST4DecodeWorker: static_cast<qsizetype> fixes for GCC 15 compatibility
- Configuration.cpp: DataLocation → AppLocalDataLocation fix
- CMake/make_icns.py: deterministic .icns generator (replaces iconutil)
- CMakeLists.txt: new decoder sources integrated into wsjt_qt_CXXSRCS
- Removed decodium_x64.iss (replaced by installer/decodium_setup.iss)
- build_and_package.bat: fixed cmake path and ISS reference

+++ Detector/FT2DecodeWorker.cpp
@@ -194,7 +194,7 @@ void FT2DecodeWorker::decodeAsync (AsyncDecodeRequest const& request)
   QMutexLocker runtime_lock {&decodium::fortran::runtime_mutex ()};
 
   short int iwave[kFt2AsyncSampleCount] {};
-  int const copyCount = std::min (request.audio.size (), kFt2AsyncSampleCount);
+  int const copyCount = std::min (static_cast<int>(request.audio.size ()), static_cast<int>(kFt2AsyncSampleCount));
   if (copyCount > 0)
     {
       std::copy_n (request.audio.constBegin (), copyCount, iwave);
@@ -250,7 +250,7 @@ void FT2DecodeWorker::decode (DecodeRequest const& request)
     }
 
   short int iwave[kFt2AsyncSampleCount] {};
-  int const copyCount = std::min (request.audio.size (), kFt2AsyncSampleCount);
+  int const copyCount = std::min (static_cast<int>(request.audio.size ()), static_cast<int>(kFt2AsyncSampleCount));
   if (copyCount > 0)
     {
       std::copy_n (request.audio.constBegin (), copyCount, iwave);
