+++ Detector/FT8DecodeWorker.cpp
@@ -138,7 +138,7 @@ void FT8DecodeWorker::decode (DecodeRequest const& request)
   QMutexLocker runtime_lock {&decodium::fortran::runtime_mutex ()};
 
   short int iwave[kFt8SampleCount] {};
-  int const copyCount = std::min (request.audio.size (), kFt8SampleCount);
+  int const copyCount = std::min (static_cast<int>(request.audio.size ()), static_cast<int>(kFt8SampleCount));
   if (copyCount > 0)
     {
       std::copy_n (request.audio.constBegin (), copyCount, iwave);
